indexing memory issue
